Fixed #194: Made async_to_sync work inside a fork
sync_to_async was keeping a threadlocal for the main event loop that was
invalid, but still present, after a process fork. This makes
async_to_sync check for the PID for that parent loop as well as the mere
presence of it, fixing this problem and letting it make a new loop in
the subprocess.